agentflow/requirements.txt
aiohttp
requests
numpy
agentops
opentelemetry-instrumentation-openai
flask
setproctitle
anthropic==0.40.0
openai==1.75.0
together==1.5.4
# google-generativeai==0.8.5
google-genai
transformers
torch
tensordict
diskcache==5.6.3
easyocr==1.7.2
metapub==0.5.12
pillow==11.1.0
platformdirs==4.3.6
pymed==0.8.9
python-dotenv==1.0.1
sympy>=1.13.1
tenacity==9.0.0
wikipedia==1.4.0
dashscope==1.24.2
litellm>=2.1.1
# ollama==0.5.1
# e2b_code_interpreter==2.0.0
# vllm - Optional dependency for running vLLM servers
